Material Support
* added support to store material within global resources
* added function to generate descriptor sets for material bindings
* removed static loaded debug resources
* added support to bind virtual buffer and buffer views to discrete bindings
* made matching dto definition for material

+++ src/Engine.GFX/Material/MaterialMetaDto.hpp
@@ -5,9 +5,9 @@
 namespace ember::engine::gfx::experimental {
 
     struct MaterialMetaDto final {
-        u16 diffuse;
-        u16 normal;
-        u16 roughness;
-        u16 ao;
+        u32 diffuse;
+        u32 normal;
+        u32 roughness;
+        u32 ao;
     };
 }
